+++ ErsatzTV.Core/Domain/FFmpegProfileVideoFormat.cs
@@ -1,0 +1,12 @@
+﻿namespace ErsatzTV.Core.Domain;
+
+public enum FFmpegProfileVideoFormat
+{
+    None = 0,
+    
+    H264 = 1,
+    Hevc = 2,
+    Mpeg2Video = 3,
+    
+    Copy = 99
+}